wip: ui improvements
